deploy/requirements.txt
aws-cdk.core==1.49.1
aws_cdk.aws_cognito==1.49.1
aws_cdk.aws_certificatemanager==1.49.1